#226 Add the Create Service popup
