+++ client/src/modules/admin/services/index.js
@@ -1,0 +1,3 @@
+import Container from './Container';
+
+export default Container;
